Break out property list handlers into their own modules
