+++ scripts/scraper-ng/rules/html-require-recipe-ingredients/ingredient-handlers/data-permitted-properties.js
@@ -1,0 +1,13 @@
+const listSectionHandler = require("./property-lists");
+
+const handleDataPermittedProperties = listSectionHandler(
+  {
+    id: "Permitted_properties",
+    introText:
+      "Rules whose selectors include this element may only use the following CSS properties:",
+    minimumListItems: 1,
+  },
+  true
+);
+
+module.exports = handleDataPermittedProperties;
